Create frontend for reset password page
