+++ frontend-client/app/(authentication)/reset-password/resetpassword.css
@@ -1,0 +1,39 @@
+.resetpasswordpage {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+}
+
+.card {
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 5px 0 rgba(0, 0, 0, 0.19);
+  width: 500px;
+}
+
+.resetpasswordpage label,
+p {
+  color: black;
+}
+
+#reset-password {
+  background-color: #3944bc;
+  color: white;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.5);
+  padding-top: 0.6em;
+  padding-bottom: 0.6em;
+  padding-left: 0.8em;
+  padding-right: 0.8em;
+  font-size: 18px;
+  cursor: pointer;
+  margin-top: 20px;
+  width: 200px;
+  display: inline-block;
+}
+
+.suddenly-remember-link {
+  text-decoration: none;
+}
+
+.suddenly-remember {
+  margin-top: 20px;
+}
